rpc.stub class moved under test
